refactor(ui): extract fixed row label column
